Added logistic map and refactored phi matrix
