+++ ceha/compress.py
@@ -1,0 +1,62 @@
+""" Module controlling image compression and decompression as well as 
+associated algorithms.
+
+"""
+
+import logging
+import numpy as np
+
+logging.basicConfig(level=logging.DEBUG, format=' %(asctime)s - %(levelname)s - %(message)s')
+
+def logistic_map(x_0, n, mu=3.99):
+    """ Generates logistic map.
+    
+    :param float x_0: Value of x at zero.
+    :param float mu: Value of mu.
+    :param float n: Number of iterations.
+    
+    """
+
+    x = np.zeros(2*n)
+
+    x[0] = x_0
+    for i, prev in enumerate(x):
+        if i < 2*n-1:
+            x[i+1] = prev * (1 - prev)
+    return x[n:2*n]
+
+
+def phi_matrix(log_map, lamb):
+    """Generates phi matrix.
+    :param log_map The log map input.
+    :param lamb The value for lambda.
+
+    """
+
+    n = len(log_map)
+    pm = np.zeros((2*n,n))
+
+    m = 2 * n
+    pm[0,:] = log_map
+    for i in range(1,m):
+        pm[i,0] = pm[i-1,n-1] * lamb
+        pm[i,1:n] = pm[i-1,0:n-1]
+    return pm
+
+if __name__ == '__main__':
+
+    # Tests logistic map
+    #a = logistic_map(0.11, 20)
+    a = logistic_map(0.23, 20)
+
+    logging.debug('Logistic map')
+    logging.debug(a)
+    # Tests phi_matrix
+    #a = np.array([1,2,3,4])
+
+    pm = phi_matrix(a,.2)
+
+    logging.debug('Phi Matrix')
+    logging.debug(pm)
+
+    # Insert simple test cases for this module here
